refactor: move jonze client to separate package
