+++ apps/family-system/package.json
@@ -19,6 +19,7 @@
 	},
 	"dependencies": {
 		"@asu/payload-storage-box": "workspace:*",
+		"@asu/jonze": "workspace:*",
 		"@payloadcms/db-postgres": "3.0.0-beta.73",
 		"@payloadcms/email-resend": "3.0.0-beta.73",
 		"@payloadcms/next": "3.0.0-beta.73",
