+++ packages/jonze/tsconfig.json
@@ -1,0 +1,34 @@
+{
+    "compilerOptions": {
+      /* Base Options: */
+      "esModuleInterop": true,
+      "skipLibCheck": true,
+      "target": "es2022",
+      "allowJs": true,
+      "resolveJsonModule": true,
+      "moduleDetection": "force",
+      "isolatedModules": true,
+      "verbatimModuleSyntax": true,
+  
+      /* Strictness */
+      "strict": true,
+      "noUncheckedIndexedAccess": true,
+      "noImplicitOverride": true,
+  
+      /* If transpiling with TypeScript: */
+      "module": "NodeNext",
+      "outDir": "dist",
+      "sourceMap": true,
+  
+      /* AND if you're building for a library: */
+      "declaration": true,
+  
+      /* AND if you're building for a library in a monorepo: */
+      "composite": true,
+      "declarationMap": true,
+  
+      /* If your code runs in the DOM: */
+        "lib": ["es2022", "dom", "dom.iterable"],
+    }
+
+  }
